Star War/requirements.txt
black==24.8.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
idna==3.10
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pytest==8.3.3
PyYAML==6.0.2
requests==2.32.3
urllib3==2.2.3